Operators generation

+++ .ci/code_generation.py
@@ -21,11 +21,14 @@ if os.name == "posix":
 else:
     LIB_TO_GENERATE = [
         "Ans.Dpf.Native.dll",
+        "Ans.Dpf.Mechanical.dll",
         "Ans.Dpf.FEMutils.dll",
         "meshOperatorsCore.dll",
         "mapdlOperatorsCore.dll",
         "Ans.Dpf.Math.dll",
+        "Ans.Dpf.PythonPluginWrapper.dll"
         "Ans.Dpf.Hdf5.dll",
+        "Ans.Dpf.FlowDiagram.dll",
         "Ans.Dpf.LSDYNAHGP.dll",
         "Ans.Dpf.LivePost.dll",
         "Ans.Dpf.PointCloudSearch.dll",
@@ -38,6 +41,10 @@ files = glob.glob(os.path.join(TARGET_PATH, "*"))
 for f in files:
     if Path(f).stem == "specification":
         continue
+    if Path(f).name == "build.py":
+        continue
+    if Path(f).name == "operator.mustache":
+        continue
     try:
         if os.path.isdir(f):
             shutil.rmtree(f)
@@ -45,7 +52,7 @@ for f in files:
             os.remove(f)
     except:
         pass
-core.start_local_server()
+core.start_local_server(config=core.AvailableServerConfigs.LegacyGrpcServer)
 code_gen = core.Operator("python_generator")
 code_gen.connect(1, TARGET_PATH)
 for lib in LIB_TO_GENERATE:
@@ -56,5 +63,3 @@ for lib in LIB_TO_GENERATE:
         code_gen.connect(2, True)
     code_gen.run()
     time.sleep(0.1)
-
-core.SERVER.shutdown()
